2.2.1.0  fixed 课时记录按班级查询显示所有校区学员记录

+++ src/main/java/JMWJ/KaoQin/XueYuanKaoQinBean.java
@@ -13,7 +13,7 @@ public class XueYuanKaoQinBean {
 	private Date timesheettime;
 	private String status;
 	private int rating;
-	private double classhours = 1.0;
+	private float classhours = 1;
 	private String comment;
 
 	public int getId() {
@@ -72,11 +72,11 @@ public class XueYuanKaoQinBean {
 		this.rating = rating;
 	}
 
-	public double getClasshours() {
+	public float getClasshours() {
 		return classhours;
 	}
 
-	public void setClasshours(double classhours) {
+	public void setClasshours(float classhours) {
 		this.classhours = classhours;
 	}
